search component 분리 작업 진행 중
