+++ pages/search/detail.tsx
@@ -3,9 +3,10 @@ import CategoryComponent from "../../src/components/search/detail/CategoryCompon
 import SearchInitComponent from "../../src/components/search/detail/SearchInitComponent";
 import NotFoundComponent from "../../src/components/search/detail/NotFoundComponent";
 import SearchResultComponent from "../../src/components/search/detail/SearchResultComponent";
-import DetailComponent from "../../src/components/search/detail/DetailComponent";
 import Head from "next/head";
 import { useState } from "react";
+import ListBookComponent from "src/components/search/detail/ListBookComponent";
+import ListStudyComponent from "src/components/search/detail/ListStudyComponent";
 
 export default function Detail(): JSX.Element {
   const [tabSelected, setTabSelected] = useState(0);
@@ -23,9 +24,8 @@ export default function Detail(): JSX.Element {
         tabSelected={tabSelected}
         setTabSelected={setTabSelected}
       />
-      {[0, 1].includes(tabSelected) && <div>책</div>}
-      {[0, 2].includes(tabSelected) && <div>스터디</div>}
-      <DetailComponent />
+      {[0, 1].includes(tabSelected) && <ListBookComponent />}
+      {[0, 2].includes(tabSelected) && <ListStudyComponent />}
     </>
   );
 }
